arm/backtrace: fix compile warning
common/arm_backtrace_unwind.c: In function 'up_backtrace':
common/arm_backtrace_unwind.c:626:27:
warning: assignment to 'long unsigned int' from 'uint8_t (*)[]'\
{aka 'unsigned char (*)[]'} makes integer from pointer without a cast [-Wint-conversion]

  626 |           frame.stack_top = &g_intstacktop;
      |

+++ arch/arm/src/common/arm_backtrace_unwind.c
@@ -623,7 +623,7 @@ int up_backtrace(struct tcb_s *tcb,
 #  ifdef CONFIG_SMP
           frame.stack_top = arm_intstack_top();
 #  else
-          frame.stack_top = &g_intstacktop;
+          frame.stack_top = (unsigned long)&g_intstacktop;
 #  endif /* CONFIG_SMP */
 #endif /* CONFIG_ARCH_INTERRUPTSTACK > 7 */
